requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-livereload-server==0.5
djangorestframework==3.14.0
livereload==2.6.3
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
